requirements.txt
Flask==2.3.3
pytest==7.4.2
pytest-html==4.1.1
allure-pytest==2.13.2
requests==2.31.0
weasyprint==52.5